Add files via upload
This file includes the code used in the portfolio analysis conducted for the purpose of the term project at Trinity College Dublin.
It uses theoretical principles of modern finance, such as Markowitz portfolio model and combines it with Monte Carlo simulation finding the opportunity set which maximizes the Sharpe ratio. Then the exact cash allocation is determined.